requirements.txt
django~=4.2.14
djangorestframework~=3.15.2
python-dotenv~=1.0.1
psycopg2~=2.9.9
psycopg2-binary~=2.9.9
pillow~=10.4.0
black~=24.10.0
django-stubs~=5.0.4
django-filter~=24.3
coverage~=7.6.1
drf-yasg~=1.21.7
setuptools~=75.1.0
dj-rest-auth~=6.0.0
django-allauth~=65.0.2
django-dispatch~=1.0.1
requests~=2.32.3